refactor: split error classes and add UnsupportedRuntimeError
