+++ src/errors/RequestError.ts
@@ -1,0 +1,6 @@
+export class RequestError extends Error {
+  constructor(message: string) {
+    super(message)
+    this.name = 'RequestError'
+  }
+}
