+++ src/errors/index.ts
@@ -1,0 +1,3 @@
+export { RequestError } from './RequestError'
+export { HttpError } from './HttpError'
+export { TimeoutError } from './TimeoutError'
